Moving GTestImplicitFENonlinearProblem into its own class
