+++ test/src/ImplicitFENonlinearProblem_test.cpp
@@ -12,106 +12,6 @@
 
 using namespace godzilla;
 
-registerObject(GTestImplicitFENonlinearProblem);
-
-static void
-f0_u(PetscInt dim,
-     PetscInt Nf,
-     PetscInt NfAux,
-     const PetscInt uOff[],
-     const PetscInt uOff_x[],
-     const PetscScalar u[],
-     const PetscScalar u_t[],
-     const PetscScalar u_x[],
-     const PetscInt aOff[],
-     const PetscInt aOff_x[],
-     const PetscScalar a[],
-     const PetscScalar a_t[],
-     const PetscScalar a_x[],
-     PetscReal t,
-     const PetscReal x[],
-     PetscInt numConstants,
-     const PetscScalar constants[],
-     PetscScalar f0[])
-{
-    f0[0] = u_t[0];
-}
-
-static void
-f1_u(PetscInt dim,
-     PetscInt Nf,
-     PetscInt NfAux,
-     const PetscInt uOff[],
-     const PetscInt uOff_x[],
-     const PetscScalar u[],
-     const PetscScalar u_t[],
-     const PetscScalar u_x[],
-     const PetscInt aOff[],
-     const PetscInt aOff_x[],
-     const PetscScalar a[],
-     const PetscScalar a_t[],
-     const PetscScalar a_x[],
-     PetscReal t,
-     const PetscReal x[],
-     PetscInt numConstants,
-     const PetscScalar constants[],
-     PetscScalar f1[])
-{
-    PetscInt d;
-    for (d = 0; d < dim; ++d)
-        f1[d] = u_x[d];
-}
-
-static void
-g0_uu(PetscInt dim,
-      PetscInt Nf,
-      PetscInt NfAux,
-      const PetscInt uOff[],
-      const PetscInt uOff_x[],
-      const PetscScalar u[],
-      const PetscScalar u_t[],
-      const PetscScalar u_x[],
-      const PetscInt aOff[],
-      const PetscInt aOff_x[],
-      const PetscScalar a[],
-      const PetscScalar a_t[],
-      const PetscScalar a_x[],
-      PetscReal t,
-      PetscReal u_tShift,
-      const PetscReal x[],
-      PetscInt numConstants,
-      const PetscScalar constants[],
-      PetscScalar g3[])
-{
-    g3[0] = 1.0 * u_tShift;
-}
-
-static void
-g3_uu(PetscInt dim,
-      PetscInt Nf,
-      PetscInt NfAux,
-      const PetscInt uOff[],
-      const PetscInt uOff_x[],
-      const PetscScalar u[],
-      const PetscScalar u_t[],
-      const PetscScalar u_x[],
-      const PetscInt aOff[],
-      const PetscInt aOff_x[],
-      const PetscScalar a[],
-      const PetscScalar a_t[],
-      const PetscScalar a_x[],
-      PetscReal t,
-      PetscReal u_tShift,
-      const PetscReal x[],
-      PetscInt numConstants,
-      const PetscScalar constants[],
-      PetscScalar g3[])
-{
-    PetscInt d;
-    for (d = 0; d < dim; ++d)
-        g3[d * dim + d] = 1.0;
-}
-
 TEST_F(ImplicitFENonlinearProblemTest, run)
 {
     auto mesh = gMesh1d();
@@ -178,29 +78,4 @@ TEST_F(ImplicitFENonlinearProblemTest, wrong_scheme)
 
     EXPECT_THAT(testing::internal::GetCapturedStderr(),
                 testing::HasSubstr("prob: The 'scheme' parameter can be either 'beuler' or 'cn'."));
-}
-
-// GTestImplicitFENonlinearProblem
-
-GTestImplicitFENonlinearProblem::GTestImplicitFENonlinearProblem(const InputParameters & params) :
-    ImplicitFENonlinearProblem(params),
-    iu(0)
-{
-}
-
-GTestImplicitFENonlinearProblem::~GTestImplicitFENonlinearProblem() {}
-
-void
-GTestImplicitFENonlinearProblem::set_up_fields()
-{
-    _F_;
-    PetscInt order = 1;
-    add_fe(this->iu, "u", 1, order);
-}
-
-void
-GTestImplicitFENonlinearProblem::set_up_weak_form()
-{
-    set_residual_block(this->iu, f0_u, f1_u);
-    set_jacobian_block(this->iu, this->iu, g0_uu, NULL, NULL, g3_uu);
 }
